move class definition to init
